rename: bookmark.ts를 request 폴더로 이동
